requirements.txt
aniso8601==9.0.1
Babel==2.9.1
blinker==1.4
certifi==2020.12.5
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==2.0.1
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-CKEditor==0.4.6
Flask-JWT-Extended==4.2.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Moment==0.11.0
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Social==1.6.2
Flask-Testing==0.8.1
Flask-WTF==0.15.1
greenlet==1.1.0
httplib2==0.19.1
idna==3.1
importlib-metadata==4.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsmin==2.2.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
oauth2==1.9.0.post1
passlib==1.7.4
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-slugify==5.0.2
pytz==2021.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sentry-sdk==1.1.0
six==1.16.0
speaklater==1.3
text-unidecode==1.3
typing-extensions==3.10.0.0
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.5
uWSGI==2.0.19.1
visitor==0.1.3
webassets==2.0
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
